adds all of the unit tests, moves things around

+++ fabfile.py
@@ -29,9 +29,6 @@ deploy_prod, this takes the old env and the new env
 create_private_settings will create an empty private settings file in
 the appropriate place with the fields for you to fill in
 
-create_empty_env, takes in an environment name
-creates you an empty database and a virtual env
-
 The code is in
 /usr/lib/ohc/log
 
@@ -58,7 +55,7 @@ UNIX_USER = "ohc"
 DB_USER = "ohc"
 RELEASE_NAME = "elcidrfh-{branch}"
 
-VIRTUAL_ENV_PATH = "/home/{usr}/.virtualenvs/{release_name}"
+VIRTUAL_ENV_PATH = "/home/{unix_user}/.virtualenvs/{release_name}"
 PROJECT_ROOT = "/usr/lib/{unix_user}".format(unix_user=UNIX_USER)
 PROJECT_DIRECTORY = "{project_root}/{release_name}"
 BACKUP_DIR = "{project_root}/var".format(project_root=PROJECT_ROOT)
@@ -66,6 +63,7 @@ GIT_URL = "https://github.com/openhealthcare/elcid-rfh"
 
 # the daily back up
 BACKUP_NAME = "{backup_dir}/back.{dt}.{db_name}.sql"
+REMOTE_BACKUP_NAME = "{backup_dir}/live/back.{dt}.{db_name}.sql"
 
 # the release back up is take just before the release, and then restored
 RELEASE_BACKUP_NAME = "{backup_dir}/release.{dt}.{db_name}.sql"
@@ -82,8 +80,9 @@ jinja_env = Environment(loader=FileSystemLoader(TEMPLATE_DIR))
 
 
 class Env(object):
-    def __init__(self, branch):
+    def __init__(self, branch, remove_existing=False):
         self.branch = branch
+        self.remove_existing = remove_existing
 
     @property
     def project_directory(self):
@@ -99,7 +98,7 @@ class Env(object):
     @property
     def virtual_env_path(self):
         return VIRTUAL_ENV_PATH.format(
-            usr=UNIX_USER,
+            unix_user=UNIX_USER,
             release_name=self.release_name
         )
 
@@ -111,6 +110,15 @@ class Env(object):
     def backup_name(self):
         now = datetime.datetime.now()
         return BACKUP_NAME.format(
+            backup_dir=BACKUP_DIR,
+            dt=now.strftime("%d.%m.%Y"),
+            db_name=self.database_name
+        )
+
+    @property
+    def remote_backup_name(self):
+        now = datetime.datetime.now()
+        return REMOTE_BACKUP_NAME.format(
             backup_dir=BACKUP_DIR,
             dt=now.strftime("%d.%m.%Y"),
             db_name=self.database_name
@@ -131,31 +139,45 @@ def run_management_command(some_command, env):
             env.virtual_env_path, some_command
         )
         result = local(cmd, capture=True)
+        print(result.stdout)
+        print(result.stderr)
+    if result.failed:
+        raise ValueError(
+            "{} failed".format(cmd)
+        )
     return result
 
 
 def pip_create_virtual_env(new_env):
-    local("rm -rf {}".format(new_env.virtual_env_path))
+    print("Creating new environment")
+    if new_env.remove_existing:
+        local("rm -rf {}".format(new_env.virtual_env_path))
+    else:
+        if os.path.isdir(new_env.virtual_env_path):
+            raise ValueError(
+                "Directory {} already exists".format(new_env.virtual_env_path)
+            )
     local("/usr/bin/virtualenv {}".format(new_env.virtual_env_path))
     return
 
 
 def pip_install_requirements(new_env, proxy):
+    print("Installing requirements")
     pip = "{}/bin/pip".format(new_env.virtual_env_path)
-    local("{0} install --upgrade pip --proxy {1}".format(pip, proxy))
+    local("{0} install pip==9.0.1 --proxy {1}".format(pip, proxy))
     local("{0} install -r requirements.txt --proxy {1}".format(pip, proxy))
 
 
 def pip_set_project_directory(some_env):
+    print("Setting the project directory")
     local("echo '{0}' > {1}/.project".format(
         some_env.project_directory, some_env.virtual_env_path
     ))
 
 
-def postgres_command(command, capture=False):
+def postgres_command(command):
     return local(
         '{0} "{1}"'.format(DB_COMMAND_PREFIX, command),
-        capture=capture
     )
 
 
@@ -163,15 +185,24 @@ def postgres_create_database(some_env):
     """ creates a database and user if they don't already exist.
         the db_name is created from the release name
     """
+    print("Creating the database")
 
-    #  https://stackoverflow.com/questions/14549270/check-if-database-exists-in-postgresql-using-shell
-    command = "sudo -u postgres psql -lqt | cut -d \| -f 1 | grep -w {database_name} | wc -l"
-    database_exists = "0" not in local(
-        command.format(database_name=some_env.database_name), capture=True
+    select_result = local(
+        "psql -tAc \"SELECT 1 FROM pg_database WHERE datname='elcid065'\"",
     )
+    database_exists = "1" in select_result.stdout
+    print(select_result.stderr)
+    print(select_result.stdout)
 
     if database_exists:
-        postgres_command("DROP DATABASE {0}".format(some_env.database_name))
+        if some_env.remove_existing:
+            postgres_command(
+                "DROP DATABASE {0}".format(some_env.database_name)
+            )
+        else:
+            raise ValueError('database {} already exists'.format(
+                some_env.database_name
+            ))
 
     postgres_command("CREATE DATABASE {0}".format(some_env.database_name))
     postgres_command("GRANT ALL PRIVILEGES ON DATABASE {0} TO {1}".format(
@@ -180,11 +211,7 @@ def postgres_create_database(some_env):
 
 
 def postgres_load_database(backup_name, new_env):
-    if not os.path.isfile(backup_name):
-        raise ValueError("unable to find a backup at {}".format(
-            backup_name)
-        )
-
+    print("Loading the database {}".format(new_env.database_name))
     local("sudo -u postgres psql -d {0} -f {1}".format(
         new_env.database_name,
         backup_name
@@ -192,6 +219,7 @@ def postgres_load_database(backup_name, new_env):
 
 
 def services_symlink_nginx(new_env):
+    print("Symlinking nginx")
     abs_address = "{}/etc/nginx.conf".format(new_env.project_directory, "")
     if not os.path.isfile(abs_address):
         raise ValueError(
@@ -206,6 +234,7 @@ def services_symlink_nginx(new_env):
 
 
 def services_symlink_upstart(new_env):
+    print("Symlinking upstart")
     abs_address = "{}/etc/upstart.conf".format(new_env.project_directory, "")
     if not os.path.isfile(abs_address):
         raise ValueError(
@@ -221,17 +250,24 @@ def services_symlink_upstart(new_env):
 
 
 def services_create_local_settings(new_env, additional_settings):
+    print("Creating local settings")
     new_settings = copy.copy(additional_settings)
     new_settings["db_name"] = new_env.database_name
     new_settings["db_user"] = DB_USER
     template = jinja_env.get_template(
-        'conf_templates/local_settings.py.jinja2'
+        'etc/conf_templates/local_settings.py.jinja2'
     )
     output = template.render(new_settings)
     local_settings_file = '{0}/{1}/local_settings.py'.format(
         new_env.project_directory, MODULE_NAME
     )
 
+    if os.path.isfile(local_settings_file) and not new_env.remove_existing:
+        raise ValueError(
+            "local settings file {} already exists".format(
+                local_settings_file
+            )
+        )
     local("rm -f {}".format(local_settings_file))
 
     with open(local_settings_file, 'w') as f:
@@ -239,13 +275,23 @@ def services_create_local_settings(new_env, additional_settings):
 
 
 def services_create_gunicorn_conf(new_env):
-    template = jinja_env.get_template('conf_templates/gunicorn.conf.jinja2')
+    print("Creating gunicorn conf")
+    template = jinja_env.get_template(
+        'etc/conf_templates/gunicorn.conf.jinja2'
+    )
     output = template.render(
         env_name=new_env.virtual_env_path
     )
     gunicorn_conf = '{0}/etc/gunicorn.conf'.format(
         new_env.project_directory
     )
+
+    if os.path.isfile(gunicorn_conf) and not new_env.remove_existing:
+        raise ValueError(
+            'gunicorn conf {} unexpectedly already exists'.format(
+                gunicorn_conf
+            )
+        )
 
     local("rm -f {}".format(gunicorn_conf))
 
@@ -254,7 +300,8 @@ def services_create_gunicorn_conf(new_env):
 
 
 def services_create_upstart_conf(new_env):
-    template = jinja_env.get_template('conf_templates/upstart.conf.jinja2')
+    print("Creating upstart conf")
+    template = jinja_env.get_template('etc/conf_templates/upstart.conf.jinja2')
     output = template.render(
         env_name=new_env.virtual_env_path,
         project_directory=new_env.project_directory
@@ -263,6 +310,13 @@ def services_create_upstart_conf(new_env):
         new_env.project_directory
     )
 
+    if os.path.isfile(upstart_conf) and not new_env.remove_existing:
+        raise ValueError(
+            'gunicorn conf {} unexpectedly already exists'.format(
+                upstart_conf
+            )
+        )
+
     local("rm -f {}".format(upstart_conf))
 
     with open(upstart_conf, 'w') as f:
@@ -270,6 +324,7 @@ def services_create_upstart_conf(new_env):
 
 
 def restart_supervisord(new_env):
+    print("Restarting supervisord")
     # warn only in case nothing is running
     with settings(warn_only=True):
         local("pkill super; pkill gunic; pkill gloss_flask")
@@ -281,13 +336,15 @@ def restart_supervisord(new_env):
 
 
 def restart_nginx():
-    local('sudo /etc/init.d/nginx restart')
+    print("Restarting nginx")
+    local('sudo service nginx restart')
 
 
 def cron_write_backup(new_env):
     """ Creates a cron job that has a postgres user writing to a file
     """
-    template = jinja_env.get_template('conf_templates/cron_backup.jinja2')
+    print("Writing cron backup")
+    template = jinja_env.get_template('etc/conf_templates/cron_backup.jinja2')
     output = template.render(
         database_name=new_env.database_name,
         backup_dir=BACKUP_DIR
@@ -302,7 +359,8 @@ def cron_write_backup(new_env):
 def cron_copy_backup(new_env):
     """ Creates a cron job that copies a file to a remote server
     """
-    template = jinja_env.get_template('conf_templates/cron_copy.jinja2')
+    print("Writing cron copy")
+    template = jinja_env.get_template('etc/conf_templates/cron_copy.jinja2')
     fabfile = os.path.abspath(__file__).rstrip("c")  # pycs won't cut it
     output = template.render(
         fabric_file=fabfile,
@@ -317,6 +375,7 @@ def cron_copy_backup(new_env):
 
 
 def send_error_email(error, some_env):
+    print("Sending error email")
     run_management_command(
         "error_emailer '{}'".format(
             error
@@ -330,7 +389,7 @@ def copy_backup(branch_name):
     current_env = Env(branch_name)
     private_settings = get_private_settings()
     env.host_string = private_settings["host_string"]
-    env.password = private_settings["password"]
+    env.password = private_settings["remote_password"]
 
     if not os.path.isfile(current_env.backup_name):
         send_error_email(
@@ -343,13 +402,13 @@ def copy_backup(branch_name):
         with settings(warn_only=True):
             failed = put(
                 local_path=current_env.backup_name,
-                remote_path=current_env.backup_name
+                remote_path=current_env.remote_backup_name
             ).failed
 
         if failed:
             send_error_email(
                 "unable to copy backup {}".format(
-                    current_env.backup_name
+                    current_env.backup_name,
                 ),
                 current_env
             )
@@ -365,17 +424,19 @@ def get_private_settings():
 
     with open(PRIVATE_SETTINGS) as privado:
         result = json.load(privado)
-        err_template = "we require {} in your private settings"
+        err_template = "we require '{}' in your private settings"
         if "db_password" not in result:
-            raise ValueError(err_template.format("a db password"))
+            raise ValueError(err_template.format("db_password"))
         if "additional_settings" not in result:
             raise ValueError(err_template.format(
-                "an additional_settings dict (even if its empty)"
+                "additional_settings dict (even if its empty)"
             ))
         if "proxy" not in result:
-            raise ValueError(err_template.format("proxy variable"))
+            raise ValueError(err_template.format("proxy"))
+        if "remote_password" not in result:
+            raise ValueError(err_template.format("remote_password"))
         if "host_string" not in result:
-            e = "we host string to be set, this should be 127.0.0.1 on test, \
+            e = "we expect host string to be set, this should be 127.0.0.1 on test, \
 or the address you want to sync to on prod in your private settings"
             raise ValueError(e)
 
@@ -385,6 +446,13 @@ or the address you want to sync to on prod in your private settings"
 @task
 def clone_branch(branch_name):
     branch_env = Env(branch_name)
+    if os.path.isdir(branch_env.project_directory):
+        raise ValueError("{} already exists".format(
+            branch_env.project_directory
+        ))
+
+    print("Cloning into {}".format(branch_env.project_directory))
+
     local(
         "git clone -b {0} {1} {2}".format(
             branch_name,
@@ -394,14 +462,41 @@ def clone_branch(branch_name):
     )
 
 
+def diff_status(old_status_json, new_status_json):
+    old_status = json.loads(old_status_json)
+    new_status = json.loads(new_status_json)
+
+    for time_period in ["all_time", "last_week"]:
+        missing = list(set(new_status.keys() - old_status.keys()))
+
+        for m in missing:
+            print("missing {} from old".format(m))
+
+        for k, v in old_status[time_period].items():
+            if k not in new_status[time_period]:
+                print("missing {} from new".format(k))
+            new_result = new_status[time_period][k]
+            if not v == new_result:
+                print(
+                    "for {} we used to have {} but now have {}".format(
+                        k, v, new_result
+                    )
+                )
+
+
 @task
 def create_private_settings():
+    if os.path.isfile(PRIVATE_SETTINGS):
+        raise ValueError(
+            'private settings already exist at {}'.format(PRIVATE_SETTINGS)
+        )
     with open(PRIVATE_SETTINGS, "w") as privado:
         json.dump(
             dict(
                 proxy="",
                 db_password="",
                 host_string="",
+                remote_password="",
                 additional_settings={}
             ),
             privado,
@@ -409,17 +504,12 @@ def create_private_settings():
         )
 
 
-@task
-def create_empty_env(new_branch):
-    _deploy(new_branch)
-
-
-def _deploy(new_branch, backup_name=None):
+def _deploy(new_branch, backup_name=None, remove_existing=False):
     if backup_name and not os.path.isfile(backup_name):
         raise ValueError("unable to find backup {}".format(backup_name))
 
     # the new env that is going to be live
-    new_env = Env(new_branch)
+    new_env = Env(new_branch, remove_existing=remove_existing)
 
     # the private settings
     private_settings = get_private_settings()
@@ -459,7 +549,7 @@ def _deploy(new_branch, backup_name=None):
 
 def infer_current_branch():
     current_dir = os.path.abspath(os.path.dirname(__file__))
-    project_beginning = Env('').project_directory
+    project_beginning = Env('', False).project_directory
 
     if not current_dir.startswith(project_beginning):
         er_temp = 'we are in {0} but expect to be in a directory beginning \
@@ -471,7 +561,7 @@ with {1}'
 @task
 def deploy_test(backup_name):
     new_branch = infer_current_branch()
-    _deploy(new_branch, backup_name)
+    _deploy(new_branch, backup_name, remove_existing=True)
     new_status = run_management_command("status_report", Env(new_branch))
     print("=" * 20)
     print("new environment was")
@@ -486,7 +576,7 @@ def validate_private_settings():
             'we need a host string inorder to scp data to a backup server'
         )
 
-    if "password" not in private_settings:
+    if "remote_password" not in private_settings:
         raise ValueError(
             'we need the password of the backup server inorder to scp data to \
 a backup server'
@@ -533,7 +623,7 @@ def deploy_prod(old_branch):
     cron_write_backup(new_env)
     cron_copy_backup(new_env)
     old_status = run_management_command("status_report", old_env)
-    _deploy(new_branch, old_env.release_backup_name)
+    _deploy(new_branch, old_env.release_backup_name, remove_existing=False)
     new_status = run_management_command("status_report", new_env)
 
     print("=" * 20)
